+++ ontology/ontology-server/src/main/java/com/aircas/ptr/foundry/ontology/entity/bo/OntologyPropertyBO.java
@@ -79,5 +79,5 @@ public class OntologyPropertyBO {
     /**
      * 数据源ID
      */
-    private Long datasourceId;
+    private int datasourceId;
 }
